+++ packages/client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/client
 
+## 13.4.17
+
+### Patch Changes
+
+- Updated dependencies [[`79fbf1c95f4fa9bfe6af17aa3e13cf18424fc521`](https://github.com/vercel/vercel/commit/79fbf1c95f4fa9bfe6af17aa3e13cf18424fc521)]:
+  - @vercel/error-utils@2.0.3
+
 ## 13.4.16
 
 ### Patch Changes